requirements.txt
blis==0.4.1
boto==2.49.0
boto3==1.13.1
botocore==1.16.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
docutils==0.15.2
emoji==0.5.4
future==0.18.2
gensim==3.8.2
idna==2.9
importlib-metadata==1.6.0
jmespath==0.9.5
joblib==0.14.1
kiwisolver==1.2.0
matplotlib==3.2.1
murmurhash==1.0.2
networkx==2.4
nltk==3.5
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.3
Pillow==7.1.2
pkg-resources==0.0.0
plac==1.1.3
preshed==3.0.2
pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-2.2.5/pt_core_news_sm-2.2.5.tar.gz
pymongo==3.10.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-twitter==3.5
pytz==2020.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
scipy==1.4.1
seaborn==0.10.1
six==1.14.0
smart-open==2.0.0
spacy==2.2.4
srsly==1.0.2
thinc==7.4.0
tqdm==4.46.0
urllib3==1.25.9
wasabi==0.6.0
wordcloud==1.7.0
zipp==3.1.0
